src/Advanced-RAG/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gradio>=5.0.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-qdrant>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
qdrant-client>=1.11.0
sentence-transformers>=3.1.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.0.0
pdfplumber>=0.11.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-multipart>=0.0.12
google-genai>=1.0.0
pydantic-settings>=2.5.0
